% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcctemplate.R
\docType{package}
\name{tcctemplate-package}
\alias{tcctemplate}
\alias{tcctemplate-package}
\title{tcctemplate: Rmarkdown Theme for DEST TCC documents}
\description{
A set of custom Rmarkdown templates for documents and
   presentations.
}
\details{
Trial run on RMarkdown templates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernandomayer/tcctemplate}
  \item Report bugs at \url{https://github.com/fernandomayer/tcctemplate/issues}
}

}
\author{
\strong{Maintainer}: Fernando Mayer \email{fernando.mayer@ufpr.br}

}
